Renames some files
