+++ apps/judicial-system/backend/src/app/messages/index.ts
@@ -1,7 +1,7 @@
-export { core } from './core'
-export { request } from './request'
-export { ruling } from './ruling'
-export { courtRecord } from './courtRecord'
-export { custodyNotice } from './custodyNotice'
+export { core } from './pdfCore'
+export { request } from './pdfRequest'
+export { ruling } from './pdfRuling'
+export { courtRecord } from './pdfCourtRecord'
+export { custodyNotice } from './pdfCustodyNotice'
 export { notifications } from './notifications'
 export { courtUpload } from './courtUpload'
