+++ apps/judicial-system/backend/src/app/messages/pdfCore.ts
