+++ apps/judicial-system/backend/src/app/messages/pdfCourtRecord.ts
