+++ apps/judicial-system/backend/src/app/messages/pdfRuling.ts
