+++ apps/judicial-system/backend/src/app/modules/defendant/index.ts
@@ -1,0 +1,1 @@
+export { DefendantModule } from './defendant.module'
